Reordering and gathering now working for the atomic density
